Major architecture change, PC partial support
