+++ inc/lis3dsh.h
@@ -1,0 +1,62 @@
+#ifndef _LIS3DSH_H_
+#define _LIS3DSH_H_
+
+#include <libopencm3/stm32/rcc.h>
+#include "libopencm3/stm32/memorymap.h"
+#include <libopencm3/stm32/usart.h>
+#include <libopencm3/stm32/gpio.h>
+#include <libopencm3/stm32/spi.h>
+#include <libopencm3/cm3/nvic.h>
+
+
+/* LIS3DSH registers addresses */
+#define ADD_REG_WHO_AM_I				0x0F
+#define ADD_REG_CTRL_4					0x20
+#define ADD_REG_OUT_X_L					0x28
+#define ADD_REG_OUT_X_H					0x29
+#define ADD_REG_OUT_Y_L					0x2A
+#define ADD_REG_OUT_Y_H					0x2B
+#define ADD_REG_OUT_Z_L					0x2C
+#define ADD_REG_OUT_Z_H					0x2D
+
+/* WHO AM I register default value */
+#define UC_WHO_AM_I_DEFAULT_VALUE		0x3F
+
+/* ADD_REG_CTRL_4 register configuration value:
+ * X,Y,Z axis enabled and 400Hz of output data rate */
+#define UC_ADD_REG_CTRL_4_CFG_VALUE		0x77
+
+/* Sensitivity for 2G range [mg/digit] */
+#define SENS_2G_RANGE_MG_PER_DIGIT		((float)0.06) // default : 0.06
+
+/* LED threshold value in mg */
+#define LED_TH_MG				(550)	/* 1000mg (1G) */
+
+/* ---------------- Local Macros ----------------- */
+
+/* set read single command. Attention: command must be 0x3F at most */
+#define SET_READ_SINGLE_CMD(x)			(x | 0x80)
+/* set read multiple command. Attention: command must be 0x3F at most */
+#define SET_READ_MULTI_CMD(x)			(x | 0xC0)
+/* set write single command. Attention: command must be 0x3F at most */
+#define SET_WRITE_SINGLE_CMD(x)			(x & (~(0xC0)))
+/* set write multiple command. Attention: command must be 0x3F at most */
+#define SET_WRITE_MULTI_CMD(x)			(x & (~(0x80))	\
+						 x |= 0x40)
+
+
+/* ------------- Local functions prototypes --------------- */
+
+extern void gpio_setup(void);
+extern void spi_setup(void);
+extern void lis3dsh_init(void);
+extern void lis3dsh_write_reg(int, int);
+extern int lis3dsh_read_reg(int);
+extern void send(uint16_t data);
+extern int two_compl_to_int16(int);
+
+extern void usart_setup(void);
+extern void send_string (char *s_de_char);
+extern char * custom_itoa(int i, char b[]);
+
+#endif
