+++ inc/serial_port.h
@@ -1,48 +1,13 @@
-/****************************************************************************
- *
- *   Copyright (c) 2014 MAVlink Development Team. All rights reserved.
- *   Author: Trent Lukaczyk, <aerialhedgehog@gmail.com>
- *           Jaycee Lock,    <jaycee.lock@gmail.com>
- *           Lorenz Meier,   <lm@inf.ethz.ch>
- *   
- *   Porting to STM32F4-discovery done by : 
- * 			 Mohamed Hage Hassan, <mohamed.hagehassan@yahoo.com>
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name PX4 nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- ****************************************************************************/
-
 #ifndef SERIAL_PORT_H_
 #define SERIAL_PORT_H_
 
 #include <cstdlib>
 #include <stdio.h>   // Standard input/output definitions
 #include <unistd.h>  // UNIX standard function definitions
+
+#include <common/mavlink.h>
+
+#ifdef STM32F4
 
 #include "libopencm3/cm3/common.h"	
 #include "libopencm3/stm32/gpio.h"
@@ -53,134 +18,72 @@
 #include "libopencm3/stm32/spi.h"
 #include <libopencm3/cm3/nvic.h>
 
+#ifdef __cplusplus
+extern "C" {
+#endif 
 
-#include <common/mavlink.h>
-                                                // bit number  876543210987654321
-#define MAVLINK_MSG_SET_POSITION_TARGET_LOCAL_NED_POSITION     0b0000110111111000
-#define MAVLINK_MSG_SET_POSITION_TARGET_LOCAL_NED_VELOCITY     0b0000110111000111
-#define MAVLINK_MSG_SET_POSITION_TARGET_LOCAL_NED_ACCELERATION 0b0000110000111111
-#define MAVLINK_MSG_SET_POSITION_TARGET_LOCAL_NED_FORCE        0b0000111000111111
-#define MAVLINK_MSG_SET_POSITION_TARGET_LOCAL_NED_YAW_ANGLE    0b0000100111111111
-#define MAVLINK_MSG_SET_POSITION_TARGET_LOCAL_NED_YAW_RATE     0b0000010111111111
+#include <lis3dsh.h>
 
+#ifdef __cplusplus
+}
+#endif
 
-//   Data Structures
-struct Time_Stamps{
-	Time_Stamps(){
-		reset_timestamps();
-	}
+// Time related structures
+struct timeval {
+		time_t      tv_sec;     /* seconds */
+		suseconds_t tv_usec;    /* microseconds */
+		};
+struct timezone {
+		int tz_minuteswest;     /* minutes west of Greenwich */
+		int tz_dsttime;         /* type of DST correction */
+		};
 
-	uint64_t heartbeat;
-	uint64_t sys_status;
-	uint64_t battery_status;
-	uint64_t radio_status;
-	uint64_t local_position_ned;
-	uint64_t global_position_int;
-	uint64_t position_target_local_ned;
-	uint64_t position_target_global_int;
-	uint64_t highres_imu;
-	uint64_t attitude;
+#else 
 
-	void reset_timestamps(){
-		heartbeat = 0;
-		sys_status = 0;
-		battery_status = 0;
-		radio_status = 0;
-		local_position_ned = 0;
-		global_position_int = 0;
-		position_target_local_ned = 0;
-		position_target_global_int = 0;
-		highres_imu = 0;
-		attitude = 0;
-	}
-	};
-// Struct containing information on the MAV we are currently connected to
-struct Mavlink_Messages {
+#include <sys/types.h>
+#include <sys/stat.h>
+#include <string.h>                  /* declaration of bzero() */
+#include <fcntl.h>
+#include <termios.h>
 
-	int sysid;
-	int compid;
+#include <sys/time.h>
+// MOD
+#include <time.h>
+//
 
-	// Heartbeat
-	mavlink_heartbeat_t heartbeat;
+#define GPIOA 1
+#define GPIOD 1
+#define GPIO6 1
+#define GPIO12 1
+#define GPIO13 1
+#define GPIO14 1
+#define USART3 3
 
-	// System Status
-	mavlink_sys_status_t sys_status;
+int init_rs232();
+void free_rs232();
+void sendcmd(int,char*);
+//struct termios oldtio,newtio;
 
-	// Battery Status
-	mavlink_battery_status_t battery_status;
+//#define BAUDRATE B4800
+#define BAUDRATE B57600
+#define RS232_DEVICE "/dev/ttyUSB0"
 
-	// Radio Status
-	mavlink_radio_status_t radio_status;
+int usart_recv_blocking(int i);
 
-	// Local Position
-	mavlink_local_position_ned_t local_position_ned;
+#endif 
 
-	// Global Position
-	mavlink_global_position_int_t global_position_int;
-
-	// Local Position Target
-	mavlink_position_target_local_ned_t position_target_local_ned;
-
-	// Global Position Target
-	mavlink_position_target_global_int_t position_target_global_int;
-
-	// HiRes IMU
-	mavlink_highres_imu_t highres_imu;
-
-	// Attitude
-	mavlink_attitude_t attitude;
-
-	// System Parameters?
-
-
-	// Time Stamps
-	Time_Stamps time_stamps;
-
-	void reset_timestamps(){
-		time_stamps.reset_timestamps();
-	}
-	};
-
-
-void serial_start(void);
-
-int get_time_sec(struct timeval *tv, struct timezone *tz);
-uint64_t get_time_usec(void);
 
 // Initialisation
-void autopilot_intialize(void);
-void autopilot_start(void);
+void serial_start(void);
 
-//	READ 
-
-// MOD 
-void global_read_messages(void);
-// 
-
-void read_messages(void);
-int usart_read_message(mavlink_message_t &message);
+// Read
+int serial_read_message(mavlink_message_t &message);
 
 // Write
-void autopilot_write(void);
-void autopilot_write_message(mavlink_message_t message);
-void autopilot_write_setpoint(void);
 int serial_write_message(const mavlink_message_t &message);
-void autopilot_update_setpoint(mavlink_set_position_target_local_ned_t setpoint);
 
-// Offboard Control
+// Time related
+int get_time_sec(struct timeval *tv, struct timezone *tz);
 
-void commands(void);
-void disable_offboard_control(void);
-void enable_offboard_control(void);
-int toggle_offboard_control( bool flag );
-
-// Control
-
-void set_position(float x, float y, float z, mavlink_set_position_target_local_ned_t &sp);
-
-// Scheduler
-
-void operation (float timer);
-void set__(float x, float y, float z, mavlink_set_position_target_local_ned_t &set_point);
 
 #endif
