+++ makefile.pc
@@ -1,0 +1,28 @@
+NAME = mavlink_control
+
+# Sources
+SRC = src
+
+# Includes
+INCLUDE = inc/
+
+# MAVLink library
+MAVLIB = mavlink/include/mavlink/v1.0/
+
+OBJS = $(NAME).o interface.o serial_port_pc.o
+
+all: $(NAME)
+
+$(NAME): git_submodule $(OBJS)
+	g++ $(OBJS) -o $(NAME) 
+
+$(NAME).o: $(SRC)/$(NAME).cpp
+	g++ -I$(INCLUDE) -I$(MAVLIB) -c $(SRC)/$(NAME).cpp
+	g++ -I$(INCLUDE) -I$(MAVLIB) -c $(SRC)/interface.cpp
+	g++ -I$(INCLUDE) -I$(MAVLIB) -c $(SRC)/serial_port_pc.cpp
+
+git_submodule:
+	git submodule update --init --recursive
+
+clean:
+	rm *.o $(NAME)
